refactoring of camel-cxf-jaxws-security-cdi
